refine C++ and Python tests
